NXP-28396: Remove unused code
